Move pure SVG elements to `pureDom`
