+++ DevToolkt/pureDom/src/jvmMain/kotlin/dev/toolkt/dom/pure/svg/PureSvgMarker.kt
@@ -1,4 +1,4 @@
-package diy.lingerie.simple_dom.svg
+package dev.toolkt.dom.pure.svg
 
 import dev.toolkt.geometry.Point
 import dev.toolkt.geometry.Size
