+++ app/build.gradle.kts
@@ -43,9 +43,6 @@ kotlin {
         }
 
         jvmMain.dependencies {
-            implementation(libs.batik.anim)
-            implementation(libs.batik.svg.dom)
-            implementation(libs.batik.css)
             implementation(libs.fop)
             implementation(libs.clikt)
         }
