+++ app/src/jvmMain/kotlin/diy/lingerie/pattern_tool/PatternPiecePreparator.kt
@@ -1,6 +1,6 @@
 package diy.lingerie.pattern_tool
 
-import diy.lingerie.simple_dom.svg.PureSvgRoot
+import dev.toolkt.dom.pure.svg.PureSvgRoot
 
 abstract class PatternPiecePreparator {
     abstract fun preparePatternPieceOutlines(
