+++ app/src/jvmMain/kotlin/diy/lingerie/pattern_tool/main.kt
@@ -9,7 +9,7 @@ import com.github.ajalt.clikt.parameters.options.option
 import com.github.ajalt.clikt.parameters.options.required
 import com.github.ajalt.clikt.parameters.types.enum
 import com.github.ajalt.clikt.parameters.types.path
-import diy.lingerie.simple_dom.svg.PureSvgRoot
+import dev.toolkt.dom.pure.svg.PureSvgRoot
 import java.nio.file.Path
 
 enum class PatternPieceId {
