+++ app/src/jvmMain/kotlin/diy/lingerie/stat_tool/main.kt
@@ -6,11 +6,11 @@ import com.github.ajalt.clikt.parameters.arguments.argument
 import com.github.ajalt.clikt.parameters.arguments.help
 import com.github.ajalt.clikt.parameters.types.path
 import dev.toolkt.dom.pure.PureColor
+import dev.toolkt.dom.pure.svg.PureSvgRoot
+import dev.toolkt.dom.pure.svg.parse
 import dev.toolkt.geometry.curves.BezierCurve
 import dev.toolkt.geometry.curves.OpenCurve
 import dev.toolkt.geometry.splines.OpenSpline
-import diy.lingerie.simple_dom.svg.PureSvgRoot
-import diy.lingerie.simple_dom.svg.parse
 import diy.lingerie.tool_utils.Playground
 import diy.lingerie.tool_utils.RecognizedShape
 import java.nio.file.Path
