+++ gradle/libs.versions.toml
@@ -1,15 +1,11 @@
 [versions]
 kotlinx-benchmark = "0.4.13"
 kotlin = "2.2.0-Beta1"
-batik = "1.19"
 kotlinx-serialization = "1.6.0"
 
 [libraries]
 kotlin-test = { module = "org.jetbrains.kotlin:kotlin-test", version.ref = "kotlin" }
 kotlinx-benchmark-runtime = { module = "org.jetbrains.kotlinx:kotlinx-benchmark-runtime", version.ref = "kotlinx-benchmark" }
-batik-anim = { module = "org.apache.xmlgraphics:batik-anim", version.ref = "batik" }
-batik-svg-dom = { module = "org.apache.xmlgraphics:batik-svg-dom", version.ref = "batik" }
-batik-css = { module = "org.apache.xmlgraphics:batik-css", version.ref = "batik" }
 fop = { module = "org.apache.xmlgraphics:fop", version = "2.10" }
 clikt = { module = "com.github.ajalt.clikt:clikt", version = "5.0.3" }
 kotlinx-serialization-json = { module = "org.jetbrains.kotlinx:kotlinx-serialization-json", version.ref = "kotlinx-serialization" }
